Move sentry; move downtime notifications
