+++ src/js/health-records/containers/HealthRecordsApp.jsx
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import { connect } from 'react-redux';
 
-import DowntimeNotification, { services } from '../../common/containers/DowntimeNotification';
+import DowntimeNotification, { services } from '../../../platform/monitoring/DowntimeNotification';
 import Modal from '../../common/components/Modal';
 import MHVApp from '../../common/containers/MHVApp';
 import RequiredLoginView from '../../common/components/RequiredLoginView';
