+++ src/js/messaging/containers/MessagingApp.jsx
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import { connect } from 'react-redux';
 
-import DowntimeNotification, { services } from '../../common/containers/DowntimeNotification';
+import DowntimeNotification, { services } from '../../../platform/monitoring/DowntimeNotification';
 import MHVApp from '../../common/containers/MHVApp';
 import AlertBox from '@department-of-veterans-affairs/jean-pants/AlertBox';
 import RequiredLoginView from '../../common/components/RequiredLoginView';
