+++ src/js/personalization/account-beta/containers/AccountApp.jsx
@@ -6,7 +6,7 @@ import { fetchLatestTerms, acceptTerms } from '../actions';
 
 import AccountMain from '../components/AccountMain';
 import RequiredLoginView from '../../../common/components/RequiredLoginView';
-import DowntimeNotification, { services } from '../../../common/containers/DowntimeNotification';
+import DowntimeNotification, { services } from '../../../../platform/monitoring/DowntimeNotification';
 import BetaApp, { features } from '../../beta-enrollment/containers/BetaApp';
 
 class UserProfileApp extends React.Component {
