+++ src/js/personalization/va-profile-beta/containers/VAProfileApp.jsx
@@ -11,7 +11,7 @@ import {
 
 import BetaApp, { features } from '../../beta-enrollment/containers/BetaApp';
 import RequiredLoginView from '../../../common/components/RequiredLoginView';
-import DowntimeNotification, { services } from '../../../common/containers/DowntimeNotification';
+import DowntimeNotification, { services } from '../../../../platform/monitoring/DowntimeNotification';
 import ProfileView from '../components/ProfileView';
 
 class UserProfileApp extends React.Component {
