+++ src/platform/monitoring/DowntimeNotification.jsx
@@ -1,11 +1,12 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import moment from '../utils/moment-setup';
 import objectValues from 'lodash/fp/values';
 import { connect } from 'react-redux';
-import { getScheduledDowntime } from '../actions';
-import Modal from '../components/Modal';
-import LoadingIndicator from '../components/LoadingIndicator';
+
+import { getScheduledDowntime } from '../../js/common/actions';
+import moment from '../../js/common/utils/moment-setup';
+import Modal from '../../js/common/components/Modal';
+import LoadingIndicator from '../../js/common/components/LoadingIndicator';
 
 export const services = {
   appeals: 'appeals',
